AFDP-6177
Former-commit-id: 33d8281df2908d9edaa20f8e64e4006088288b81

+++ acm-services/acm-service-ecm/src/main/java/com/armedia/acm/plugins/ecm/service/lock/ContainerLockingProvider.java
@@ -66,7 +66,8 @@ public class ContainerLockingProvider implements ObjectLockingProvider
         AcmObjectLock existingLock = objectLockService.findLock(objectId, objectType);
 
         Date now = new Date(System.currentTimeMillis());
-        if (now.after(existingLock.getExpiry())) {
+        if (existingLock != null && now.after(existingLock.getExpiry()))
+        {
             //lock has expired and will be removed
             objectLockService.removeLock(existingLock);
             existingLock = null;
